Fold action-rejection.ts into overseer.ts
Per review feedback on #203: avoid a new single-function file. The
rejection core moves into overseer.ts unchanged (module-level, still
injected/testable); a combined actions.ts split can follow when
overseer.ts is broken up more deliberately.
